requirements.txt
yfinance
pandas
openai
python-dotenv
termcolor
tqdm
alpaca-py
ddgs
supabase
numpy
